+++ src/HealthCheckServiceProvider.php
@@ -14,6 +14,10 @@ class HealthCheckServiceProvider extends ServiceProvider
     public function boot()
     {
         $this->loadRoutesFrom(__DIR__.'/routes.php');
+
+        $this->publishes([
+            __DIR__.'/config/healthcheck.php' => config_path('healthcheck.php'),
+        ]);
     }
 
     /**