+++ src/config/healthcheck.php
@@ -1,0 +1,11 @@
+<?php
+
+/*
+ * Custom properties for the health check package
+ *
+ */
+return [
+    'ssl' => true,
+    'database' => true,
+    'application' => true,
+];